Added quicksave, added masking, cleanup
