+++ i18n/com/romraider/swing/ECUEditorMenuBar.properties
@@ -1,6 +1,8 @@
 FILE = File
 OPENIMG = Open Image ...
 OPENIMGS = Open Image(s) ...
+SAVE = Save Image ...
+QUICKSAVE = Save {0}
 SAVEAS = Save Image As ...
 SAVEFAS = Save {0} As ...
 SAVEREPO = Save Image As Repository ...
