+++ src/main/java/com/romraider/util/ByteUtil.java
@@ -159,5 +159,34 @@ public final class ByteUtil {
         return failure;
     }
     
-    
+	//Java 9 Method
+	public static int parseUnsignedInt(String s, int radix) throws NumberFormatException {
+	    if (s == null)  {
+	        throw new NumberFormatException("null");
+	    }
+	
+	    int len = s.length();
+	    if (len > 0) {
+	        char firstChar = s.charAt(0);
+	        if (firstChar == '-') {
+	            throw new
+	                NumberFormatException(String.format("Illegal leading minus sign " +
+	                                                   "on unsigned string %s.", s));
+	        } else {
+	            if (len <= 5 ||(radix == 10 && len <= 9) ) {
+	                return Integer.parseInt(s, radix);
+	            } else {
+	                long ell = Long.parseLong(s, radix);
+	                if ((ell & 0xffffffff00000000L) == 0) {
+	                    return (int) ell;
+	                } else {
+	                    throw new
+	                        NumberFormatException(String.format("String value %s exceeds " + "range of unsigned int.", s));
+	                }
+	            }
+	        }
+	    } else {
+	    	throw new NumberFormatException(s);
+	    }
+	}  
 }
